pyproject.toml
[tool.poetry]
name = "brewmingle"
version = "0.1.0"
description = ""
authors = ["DarthZiel <Studerzhan@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.0.3"
pillow = "^10.2.0"
djangorestframework = "^3.14.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
